refactor: postList 컴포넌트, postSearch 컴포넌트 ts로 마이그레이션
