+++ src/apis/search.ts
@@ -2,7 +2,7 @@ import { POST_URL } from '../utils/constants/post';
 import { baseRequest } from './core';
 export const searchAll = async (keyword: string) => {
   try {
-    const data = await baseRequest.get(POST_URL.SEARCH_POST + keyword);
+    const { data } = await baseRequest.get(POST_URL.SEARCH_POST + keyword);
     return data;
   } catch (error) {
     console.error(error);
