+++ src/components/Post/PostSearch/style.ts
